Duy backend dev

+++ backend/app/routers/playlist.py
@@ -16,12 +16,16 @@ from utils.exceptions import BAD_REQUEST
 router = APIRouter(tags=["Playlist"], prefix="/playlist")
 
 
-@router.post("", description="Create playlist")
+@router.post("",)
 async def create_playlist(
     supabase: Annotated[Client, Depends(get_supabase)],
     user_id: Annotated[str, Security(get_id)],
     playlist: BasePlaylist,
 ):
+    """
+    Create playlist
+    - **name**: each playlist must have a name
+    """ 
     try:
         supabase.table("playlist").insert(
             {"name": playlist.name, "song_id" : [], "user_id": user_id}
@@ -31,21 +35,30 @@ async def create_playlist(
         raise BAD_REQUEST
     
 
-@router.delete("", description="Delete playlist")
+@router.delete("",)
 async def delete_playlist(
     supabase: Annotated[Client, Depends(get_supabase)],
     user_id: Annotated[str, Security(get_id)],
     playlist_id: int,
 ):
+    """
+    Delete playlist
+    - **playlist_id**: id of playlist you want to delete
+    """ 
+    if supabase.table("playlist").select("*").match({"id" : playlist_id,"user_id": user_id}).execute().data ==[]:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This playlist isn't manage by you",
+        )
     try:
         supabase.table("playlist").delete().eq("id", playlist_id).execute()
-        return {"detail": "Album deleted"}
+        return {"detail": "Playlist deleted"}
     except:
         raise BAD_REQUEST
 
 # get list of all your playlist      
 @router.get("/all", description="Get list of your playlist", response_model=List[Playlist])
-async def get_all_album(
+async def get_all_playlist(
     supabase: Annotated[Client, Depends(get_supabase)],
     user_id: Annotated[str, Security(get_id)],
 ) -> List[Playlist]:
@@ -57,25 +70,29 @@ async def get_all_album(
         raise BAD_REQUEST
 
 # modify playlist
-@router.patch("", description="Modify playlist")
+@router.patch("/{playlist_id}", )
 async def modify_playlist(
     supabase: Annotated[Client, Depends(get_supabase)],
     user_id: Annotated[str, Security(get_id)],
-    playlist: Playlist,
+    playlist_id: int,
+    playlist: BasePlaylist,
     songs: List[int],
-):
+):    
+    """
+    Modify playlist
+    - **playlist_id**: id of playlist you want to update
+    - **name**: new name of playlist
+    - **songs**: the new list song of playlist
+    """ 
+    if supabase.table("playlist").select("*").match({"user_id" : user_id,"id": playlist_id}).execute().data ==[]:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This playlist isn't manage by you or not exist",
+        )
     
-    if playlist.user_id != user_id:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="You are not the owner of this playlist",
-            )
     try:
-        supabase.table("playlist").update(
-            jsonable_encoder(playlist, exclude={"id", "user_id", "created_at"})
-        ).eq("id", playlist.id).execute()
-
-        supabase.table("playlist").update({ "song_id": songs }).eq("id", playlist.id).execute()
+        supabase.table("playlist").update( { "name": playlist.name}).eq("id", playlist_id).execute()
+        supabase.table("playlist").update({ "song_id": songs }).eq("id", playlist_id).execute()
 
         return {"detail": "Playlist modified"}
     except:
@@ -85,20 +102,26 @@ async def modify_playlist(
 
 
 # get playlist and its songs
-@router.get("/{playlist_id}", description="Get playlist and its songs")
+@router.get("/{playlist_id}",)
 async def get_playlist(
     supabase: Annotated[Client, Depends(get_supabase)],
     user_id: Annotated[str, Security(get_id)],
     playlist_id: int,
 ):
+    """
+    Get detail playlist and its songs
+    - **playlist_id**: id of playlist you want to get detail
+    """ 
     try:
         res = (
             supabase.table("playlist")
-            .select("*")
+            .select("id,name,created_at,song_id")
             .eq("id", playlist_id)
             .execute()
             .dict()["data"]
         )
+        for i in range( len(res[0]["song_id"])):
+            res[0]["song_id"][i] = supabase.table("songs").select("*").eq("id", res[0]["song_id"][i]).execute().dict()["data"]
         return res
     except:
         raise BAD_REQUEST
